🎨 use colorful logger in watcher tests
