+++ modules/test-runner/src/util/helpers/fundChannel.ts
@@ -4,17 +4,17 @@ import {
   EventNames,
   IConnextClient,
 } from "@connext/types";
-import { delay, getAddressFromAssetId } from "@connext/utils";
+import { ColorfulLogger, delay, getAddressFromAssetId } from "@connext/utils";
 import { BigNumber } from "ethers/utils";
 
-import { env, expect, Logger } from "../";
+import { env, expect } from "../";
 
 export const fundChannel = async (
   client: IConnextClient,
   amount: BigNumber,
   assetId: AssetId = CONVENTION_FOR_ETH_ASSET_ID,
 ): Promise<void> => {
-  const log = new Logger("FundChannel", env.logLevel);
+  const log = new ColorfulLogger("FundChannel", env.logLevel);
   const tokenAddress = getAddressFromAssetId(assetId);
   const prevFreeBalance = await client.getFreeBalance(tokenAddress);
   await new Promise(async (resolve, reject) => {
@@ -49,7 +49,7 @@ export const requestCollateral = async (
   assetId: AssetId = CONVENTION_FOR_ETH_ASSET_ID,
   enforce: boolean = false,
 ): Promise<void> => {
-  const log = new Logger("RequestCollateral", env.logLevel);
+  const log = new ColorfulLogger("RequestCollateral", env.logLevel);
   const tokenAddress = getAddressFromAssetId(assetId);
   const preCollateralBal = await client.getFreeBalance(tokenAddress);
