+++ modules/test-runner/src/util/helpers/withdrawFromChannel.ts
@@ -1,11 +1,11 @@
 import { IConnextClient } from "@connext/types";
+import { ColorfulLogger } from "@connext/utils";
 import { Contract, Wallet } from "ethers";
 import { AddressZero } from "ethers/constants";
 import { BigNumber } from "ethers/utils";
 import tokenAbi from "human-standard-token-abi";
 
 import { env } from "../env";
-import { Logger } from "../logger";
 import { expect } from "../";
 import { ethProvider } from "../ethprovider";
 
@@ -18,7 +18,7 @@ export const withdrawFromChannel = async (
   // try to withdraw
   const preWithdrawalBalance = await client.getFreeBalance(assetId);
   const expected = preWithdrawalBalance[client.signerAddress].sub(amount);
-  const log = new Logger("WithdrawFromChannel", env.logLevel);
+  const log = new ColorfulLogger("WithdrawFromChannel", env.logLevel);
   log.info(`client.withdraw() called`);
   const start = Date.now();
   const { transaction } = await client.withdraw({
