+++ modules/test-runner/src/util/index.ts
@@ -15,7 +15,6 @@ export * from "./constants";
 export * from "./db";
 export * from "./env";
 export * from "./ethprovider";
-export * from "./logger";
 export * from "./messaging";
 export * from "./misc";
 export * from "./nats";
